tests: fix issues with the removal order test case
The test checking that the GPIO directory of a line exposed by a chip is
removed without unexporting it when its parent chip is removed has some
issues. The path we checked was wrong, the name could be more indicative
of its purpose and the gpio variable should be explicitly deleted.

Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>