gpiosim: wait for a bit for the chip's class directory to appear
As this set of tests is aimed at ensuring compatiblity of user-space
implementations with the in-kernel sysfs interface, we need to make some
concessions. One is: the user-space may lag a bit behind the kernel on
creating class directories for new chips as it will most likely rely on
uevents to be notified about it. Wait for up to a second for the new
chip to appear before calling it quits.

Signed-off-by: Bartosz Golaszewski <bartosz.golaszewski@linaro.org>